requirements.txt
fastapi
uvicorn
lxml
psycopg[binary]
pydantic
groq
python-multipart
python-dotenv
